Remove go code.
I'm going to focus on a single language, C, to avoid duplication and to keep a
focus on the algorithms rather than the implementation language.

I've chosen C since not only is that the language I use most at the moment, it's
the easiest one to port basically anywhere. I still love you though go!